Introduce full screen mode.
Firstly, the game will start in windowed mode, but it will scale it
by the largest possible integer (2, 3, 4, ...). The style of the game
is low-res: 640 x 360 (in Main.kt, it is gameResX, gameResY).
The user can toggle between fullscreen or windowed mode.
If the screen ratio doesn't match 16:9, probably something will be
drawn 'outside' the image, and then it might be visible.